Add support for Python 2.7.9 and above
Due to the following change in Python:
- https://bugs.python.org/issue22417
- https://hg.python.org/cpython/rev/fb83916c3ea1

das_client.py is failing with or without key based authentication:

    urllib2.URLError: <urlopen error [SSL: CERTIFICATE_VERIFY_FAILED]
    certificate verify failed (_ssl.c:591)>

The patch makes get_data and keys_attrs to use HTTPSHandler instead of
HTTPHandler as all calls are done over https. Then if we are running on
Python 2.7.9 or above set context to ssl._create_unverified_context() as
described in Python commit to use the old behavior.

We should figure out why cmsweb certificate does not pass the
verification step. Maybe 'CERN Grid Certification Authority' is not part
of CA bundle (e.g. it's not listed in /etc/ssl/certs/ca-bundle.crt)

Tested with the following commands:

    das_client.py --query 'file dataset=/MinimumBias/Run2011A-v1/RAW run=165121 site=T2_CH_CERN'
    voms-proxy-init -voms cms -rfc
    das_client.py --cert /tmp/x509up_u${UID} --key /tmp/x509up_u${UID} --query 'file dataset=/MinimumBias/Run2011A-v1/RAW run=165121 site=T2_CH_CERN'

Tested-on: Python 2.6.6 (SLC6)
Tested-on: Python 2.7.5 (SLC6, scl enable python27 -- bash)
Tested-on: Python 2.7.11 (SLC6, CMSSW, slc6_amd64_gcc530)
Tested-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>
Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>